Completed Pie Chart Component
Converted key entirely to Swing JComponent, it now can be organised in various ways. Pie Chart component perfectly splits between the pie chart and the key, this can also be altered and swapped around. Changed unknown extensions to "File" in the key and hid empty files from the pie chart and key.