Updated the File Explorer
The File Explorer is closer to working as intended, all the buttons work along with a new table model for viewing Drives on the computer. The right click context menu has been modified so that it doesn't the row to be selected first, this is also true for the double click. The default file table model now has code to support the user's history through the File Explorer which the back and next buttons are able to take advantage of.