Search Frame Update
This fixes the issue of the search internal frame not showing any sort of results when being searched, this had been fixed by adding a tabbed panel to the File Explorer and the search results being added as a tab. Still a few bugs remain as the tabs cannot be removed.