Update Search Pack
This update removes the placeholder internal frame and adds a new one that uses File Explorer functionality. Added a new way of searching (both; which is basically looking for both files and directories) and added more information that can be passed in the Search Event Listener. Search Frame has also been added to the GUI.