General improvement Modification (#6)
* Removed unnessary getType() function

Removed the getType function that was used to tell the difference between the file tabel models, not instanceof is used instead.

* Moved CreateFolder and improved it

Removed its useless package and moved it into the main package. Also made it so you can create folders anywhere, not just in drive A.

* Update ExplorerButton.java

Removed the code that made the rollover effect for the super class's method of rollover.

* Context Menu Update

This moves the CreateFolder method from being activated by the menu bar to the context menu whilst also integrating the file explorer into how it decides where to put the new folder. A slight effiency improvement and a dialog indicating when a File has been deleted.

* Search Improvement

This fixes the errors that the file iterator has been having along with implementing thread based searches that will use a new thread to conduct the search.

* File Explorer Improvement

This heavily improves the visual look of the file explorer whilst also increasing the efficiency of how the explorer pulled filesystem icons. FileTable is also set up to eventually support selection colours and hover colours for further improvement of visuals.